+++ src/main/scala/edu/ie3/simona/sim/setup/SimonaStandaloneSetup.scala
@@ -210,7 +210,7 @@ class SimonaStandaloneSetup(
     setupExtSim(extLinks, args)(
       context,
       scheduler,
-      simonaConfig,
+      simonaConfig.simona.powerflow.resolution,
     )
   }
 
@@ -262,8 +262,7 @@ class SimonaStandaloneSetup(
       )
 
   override def resultEventListener(
-      context: ActorContext[_],
-      extSimSetupData: ExtSimSetupData,
+      context: ActorContext[_]
   ): Seq[ActorRef[ResultEventListener.Request]] = {
     // append ResultEventListener as well to write raw output files
     Seq(