+++ src/test/scala/edu/ie3/simona/model/thermal/ThermalGridWithStorageOnlySpec.scala
@@ -40,6 +40,7 @@ class ThermalGridWithStorageOnlySpec
           thermalBusInput,
           Set.empty[ThermalHouseInput].asJava,
           Set[ThermalStorageInput](thermalStorageInput).asJava,
+          Set.empty[ThermalStorageInput].asJava,
         )
 
       ThermalGrid(thermalGridInput) match {
@@ -57,6 +58,7 @@ class ThermalGridWithStorageOnlySpec
         thermalBusInput,
         Set.empty[ThermalHouseInput].asJava,
         Set[ThermalStorageInput](thermalStorageInput).asJava,
+        Set.empty[ThermalStorageInput].asJava,
       )
     )